requirements.txt
hyperopt==0.2.7
lightgbm==4.6.0
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
pyyaml==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.1
statsmodels==0.14.4